Improve student handlers
