+++ student/get_student.go
@@ -25,3 +25,16 @@ func getStudentHandler(ctx *gin.Context) {
 
 	ctx.JSON(http.StatusOK, gin.H{"data": student})
 }
+
+func getAllStudentsHandler(ctx *gin.Context) {
+	var students []Student
+
+	err := getAllStudents(ctx, &students)
+
+	if err != nil {
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	ctx.JSON(http.StatusOK, gin.H{"data": students})
+}
